@@ -1,36 +1,45 @@
 {
   "name": "express-abort",
-  "version": "1.0.2",
+  "version": "2.0.0",
   "repository": "https://github.com/dzek69/express-abort.git",
   "author": "Jacek Nowacki @dzek69 <git-public@dzek.eu>",
   "license": "MIT",
   "main": "src/index.js",
   "scripts": {
     "test": "cross-env NODE_ENV=testing mocha 'src/**/*.spec.js'",
-    "docs": "node build-scripts/docs && jsdoc src README.md -t node_modules/docdash -d ./docs -u ./tutorials && node build-scripts/docs.after",
-    "transpile": "node build-scripts/transpile && babel src -d dist --ignore spec.js",
-    "prepublishOnly": "npm run test && npm run lint && npm run docs",
+    "docs": "node build-scripts/docs && jsdoc -r src README.md -t node_modules/docdash -d ./docs -u ./tutorials -c jsdoc.json && node build-scripts/docs.after",
+    "transpile": "node build-scripts/transpile && babel src -d dist --ignore **/*.spec.js",
+    "prepublishOnly": "npm run lint && npm run test && npm run docs",
     "prepack": "npm run transpile",
-    "lint": "eslint src/**/*.js src/*.js",
+    "lint": "cross-env eslint --report-unused-disable-directives 'src/**/*.js' 'src/*.js'",
     "lint:fix": "npm run lint -- --fix"
   },
   "devDependencies": {
-    "@babel/cli": "^7.1.5",
-    "@babel/core": "^7.1.6",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.1.6",
-    "@babel/register": "^7.0.0",
-    "@dzek69/eslint-config-base": "^0.0.1",
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.5.5",
+    "@babel/polyfill": "^7.4.4",
+    "@babel/preset-env": "^7.5.5",
+    "@babel/register": "^7.5.5",
+    "@dzek69/eslint-config-base": "^1.0.1",
     "babel-plugin-rewire": "^1.2.0",
     "cross-env": "^5.2.0",
-    "docdash": "^1.0.0",
-    "eslint": "^5.9.0",
+    "docdash": "^1.1.1",
+    "eslint": "^6.2.2",
     "fs-extra": "^7.0.1",
-    "jsdoc": "^3.5.5",
-    "mocha": "^5.2.0",
+    "husky": "^3.0.4",
+    "jsdoc": "^3.6.3",
+    "mocha": "^6.2.0",
     "must": "^0.13.4"
   },
   "dependencies": {
     "better-custom-error": "^1.0.0"
+  },
+  "libraryTemplate": {
+    "version": "1.9.0"
+  },
+  "husky": {
+    "hooks": {
+      "pre-push": "yarn prepublishOnly && yarn transpile"
+    }
   }
 }
